token auth added

+++ app/blueprints/api/social_routes.py
@@ -1,11 +1,14 @@
 from . import bp as api
 from app.models import *
-from flask import make_response, request
+from flask import make_response, request, g
+from app.blueprints.auth.auth import token_auth
 
-# returns all the posts in the database
+# returns all the posts the users they are following
 @api.get('/posts')
+@token_auth.login_required()
 def get_all_posts():
-    posts = Post.query.all()
+    user = g.current_user
+    posts = user.followed_posts()
     response_list = []
     for post in posts:
         post_dict={
@@ -21,8 +24,13 @@ def get_all_posts():
 
 # returns a single post take a post id
 @api.get('/posts/<int:id>')
+@token_auth.login_required()
 def get_single_post(id):
+    user = g.current_user
     post = Post.query.get(id)
+    # check to make sure the user has access to the post
+    if not user.is_following(post.author) and not post.author.id == user.id:
+        return make_response("Cannot get a post for someone the user is not following", 403)
     if not post:
         return make_response(f"Post ID: {id} does not exist", 404)
     response_dict={
@@ -42,11 +50,14 @@ def get_single_post(id):
 
 # Creates a new post!
 @api.post('/posts')
+@token_auth.login_required()
 def post_post():
     posted_data = request.get_json()
     u = User.query.get(posted_data['user_id'])
     if not u:
         return make_response(f"User id: {posted_data['user_id']} does not exist", 404)
+    if g.current_user.id != u.id:
+        return make_response("You can only post for yourself", 403)
     post = Post(**posted_data)
     post.save()
     return make_response(f"Post id: {post.id} created", 200)
@@ -55,9 +66,12 @@ def post_post():
 # delete a post
 # {"id":3}
 @api.delete('/posts')
+@token_auth.login_required()
 def delete_post():
     posted_data = request.get_json()
     post = Post.query.get(posted_data['id'])
+    if not g.current_user.id == post.user_id:
+        return make_response("You can only delte your post", 403)
     if not post:
         return make_response(f"Post id: {posted_data['id']} does not exist", 404)
     post.delete()
@@ -70,6 +84,7 @@ def delete_post():
 #     'user_id': 3
 # }
 @api.put('/posts')
+@token_auth.login_required()
 def put_post():
     posted_data = request.get_json()
     post = Post.query.get(posted_data['id'])
@@ -78,6 +93,8 @@ def put_post():
     u = User.query.get(posted_data['user_id'])
     if not u:
         return make_response(f"User id: {posted_data['user_id']} does not exist", 404)
+    if g.current_user.id != u.id:
+        return make_response("You can only modify your own posts", 403)
     post.user_id = posted_data['user_id']
     post.body = posted_data['body']
     post.save()
